Refactor: remove unnecessary action, refactor safe removal (#2428)
* remove unnecessary action, refactor safe removal

* move types

* remove default safe check in remove safe reducer

* fix middleware for saving default safe
