+++ src/components/SafeListSidebar/SafeList/AddressWrapper.tsx
@@ -8,7 +8,7 @@ import { SafeRecordWithNames } from 'src/logic/safe/store/selectors'
 import { formatAmount } from 'src/logic/tokens/utils/formatAmount'
 import { sameAddress } from 'src/logic/wallets/ethAddresses'
 import DefaultBadge from './DefaultBadge'
-import { DefaultSafe } from 'src/routes/safe/store/reducer/types/safe'
+import { DefaultSafe } from 'src/logic/safe/store/reducer/types/safe'
 import setDefaultSafe from 'src/logic/safe/store/actions/setDefaultSafe'
 import { getNetworkInfo } from 'src/config'
